agent/requirements.txt
aiohttp==3.9.5
openai==1.30.1
sendgrid==6.11.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
openpyxl==3.1.5
python-docx==1.1.2
fastapi==0.115.0
uvicorn[standard]==0.30.1
